Refactor into separate package and extract constants
